+++ src/client/package.json
@@ -64,9 +64,11 @@
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
     "font-awesome": "4.5.0",
+    "hoist-non-react-statics": "^2.5.5",
     "lodash": "3.10.1",
     "lodash.keyby": "4.6.0",
     "moment": "2.13.0",
+    "npm": "^6.1.0",
     "numeral": "1.5.3",
     "nvd3": "1.8.5",
     "prop-types": "15.6.0",
