+++ src/client/src/ui/shell/OpenFinProvider.tsx
@@ -1,29 +1,25 @@
 import * as React from 'react'
-import { EnvironmentConsumer } from '../../main'
 import { OpenFin } from '../../services'
+import { Environment, withEnvironment } from './EnvironmentProvider'
 import OpenFinChrome from './OpenFinChrome'
 
 interface OpenFinShellProps {
   openFin: OpenFin
+  environment: Environment
 }
 
-const OpenFinMainWindow: React.SFC<OpenFinShellProps> = ({ openFin, children }) => (
-  <EnvironmentConsumer>
-    {isRunningOnDesktop =>
-      isRunningOnDesktop ? (
-        <OpenFinChrome
-          openFin={openFin}
-          minimize={openFin.minimize.bind(openFin)}
-          maximize={openFin.maximize.bind(openFin)}
-          close={openFin.close.bind(openFin)}
-        >
-          {children}
-        </OpenFinChrome>
-      ) : (
-        children
-      )
-    }
-  </EnvironmentConsumer>
-)
+const OpenFinMainWindow: React.SFC<OpenFinShellProps> = ({ openFin, environment, children }) =>
+  environment.isRunningDesktop ? (
+    <OpenFinChrome
+      openFin={openFin}
+      minimize={openFin.minimize.bind(openFin)}
+      maximize={openFin.maximize.bind(openFin)}
+      close={openFin.close.bind(openFin)}
+    >
+      {children}
+    </OpenFinChrome>
+  ) : (
+    <React.Fragment>{children}</React.Fragment>
+  )
 
-export default OpenFinMainWindow
+export default withEnvironment(OpenFinMainWindow)
