+++ src/client/src/ui/tearoff/index.ts
@@ -1,1 +1,2 @@
 export { TearOff } from './TearOff'
+export { Portal } from './Portal'
